+++ game/models/game.py
@@ -48,7 +48,7 @@ class Master(models.Model):
 
 
 class Player(models.Model):
-    user = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.CASCADE)
+    user = models.OneToOneField(settings.AUTH_USER_MODEL, on_delete=models.CASCADE)
     game = models.ForeignKey(Game, on_delete=models.CASCADE)
     character = models.OneToOneField(Character, on_delete=models.CASCADE)